requirements.txt
# ==============================
# Core Backend
# ==============================
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1

# ==============================
# LLM Providers
# ==============================
openai==1.35.7
google-generativeai==0.7.2

# ==============================
# Embeddings & ML
# ==============================
sentence-transformers==3.0.1
transformers==4.42.4
torch==2.6.0

# ==============================
# Vector Database
# ==============================
qdrant-client==1.9.1

# ==============================
# Knowledge Graph (Neo4j)
# ==============================
neo4j==5.21.0

# ==============================
# Code Parsing (AST)
# ==============================
tree-sitter==0.21.1
tree-sitter-languages

# ==============================
# GitHub & HTTP
# ==============================
gitpython==3.1.43
requests==2.32.3

# ==============================
# Async & Performance
# ==============================
httpx==0.27.0
aiofiles==23.2.1

# ==============================
# Utilities
# ==============================
tqdm==4.66.4
numpy==1.26.4
pandas==2.2.2

# ==============================
# Testing
# ==============================
pytest==8.2.2

# ==============================
# Optional (Reranking)
# ==============================
scikit-learn==1.5.1

python-dotenv